Move classes to "Kotlin" package
The `org.gradle.internal.buildtree` package is not in the list
of "known" package that contain compiled Kotlin classes, so
it triggers an arch-test failure about "wrong" nullability annotations
